+++ source/AzAuth.PS/Cmdlets/Helpers.cs
@@ -18,7 +18,9 @@ public class ExistingAccounts : IArgumentCompleter
 
         try
         {
-            return CacheManager.GetAccounts(fakeBoundParameters["TokenCache"]?.ToString()).Select(a => new CompletionResult(a));
+            var cacheName = fakeBoundParameters["TokenCache"]?.ToString();
+            string rootDirectory = fakeBoundParameters["RootPath"]?.ToString() ?? CacheManager.GetCacheRootDirectory();
+            return CacheManager.GetAccounts(cacheName, rootDirectory).Select(a => new CompletionResult(a));
         }
         catch (Exception) { /* It's fine if we cannot get accounts here */ }
 
@@ -41,5 +43,28 @@ public class ValidateCertificatePathAttribute : ValidateArgumentsAttribute
         {
             throw new ArgumentException($"File '{path}' does not exist.");
         }
+    }
+}
+
+public class ExistingCaches : IArgumentCompleter
+{
+    public IEnumerable<CompletionResult> CompleteArgument(
+        string commandName,
+        string parameterName,
+        string wordToComplete,
+        CommandAst commandAst,
+        IDictionary fakeBoundParameters)
+    {
+        IEnumerable<CompletionResult> results = Enumerable.Empty<CompletionResult>();
+
+        try
+        {
+            var cacheName = fakeBoundParameters["TokenCache"]?.ToString();
+            string rootDirectory = fakeBoundParameters["RootPath"]?.ToString() ?? CacheManager.GetCacheRootDirectory();
+            return CacheManager.GetAvailableCaches($"{cacheName}*", rootDirectory).Select(a => new CompletionResult(a.Name));
+        }
+        catch (Exception) { /* It's fine if we cannot get accounts here */ }
+
+        return results;
     }
 }